REWORK
had to rework a lot of stuff. had to create new sort methods for sorting by height. which means no second intake and no comparable. The first bits of code just get the lines of text we need from args so users can put whatever they want and we still get what we need for the code. next is the disgusting if else statements. it needs refinement